breakout modeling utils and gain utils.
