--- calamity/echo.py
@@ -1,4 +1,0 @@
-
-def echo(message, verbose=True):
-    if verbose:
-        print(message)
