+++ calamity/simple_cov.py
@@ -1,7 +1,7 @@
 import tensorflow as tf
 import numpy as np
 import datetime
-from .echo import echo
+from .utils import echo
 
 
 def simple_cov_matrix(uvdata, baseline_group, ant_dly, dtype=np.float32):
